Only remove the file if it exists.

+++ test/file_setup_test.sh
@@ -4,21 +4,9 @@
 
 test_fixtures_dir="$PWD/test/fixtures"
 
-
-function print_state()
-{
-	echo -e "PWD:\t\t\t$PWD"
-	echo -e "test_fixtures_dir:\t$test_fixtures_dir"
-	echo -e "test_file:\t\t$test_file"
-	echo "test/fixtures:"
-	ls -a "$test_fixtures_dir"
-}
-
 function setUp()
 {
 	test_file="$test_fixtures_dir/file.txt"
-
-	print_state
 
 	echo ">>> Creating test file $test_file ..."
 	echo "test file" > "$test_file"
@@ -33,10 +21,10 @@ function test_file_creation()
 
 function tearDown()
 {
-	echo ">>> Removing test file $test_file ..."
-	rm "$test_file"
-
-	print_state
+	if [[ -f "$test_file" ]]; then
+		echo ">>> Removing test file $test_file ..."
+		rm "$test_file"
+	fi
 }
 
 SHUNIT_PARENT=$0 . $SHUNIT2
